requirements.txt
tweepy>=4.14.0
requests
python-dotenv
